v1/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "advanced-alchemy[cli]>=1.8.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.48",
]
